requirements.txt
anyio==4.2.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.0
Flask==3.0.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
python-telegram-bot==20.7
requests==2.31.0
sniffio==1.3.0
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1
gunicorn==21.2.0
logging==0.4.9.6